rename middleware's name from `internal` to `inner`
The internal package has the special meaning in go
(see https://golang.org/s/go14internal).
So rename it to `inner`.
